+++ src/components/top-nav/search-bar.test.tsx
@@ -66,6 +66,6 @@ describe('SearchBar Component', () => {
 
     expect(input.className).not.toContain('outline-none');
     expect(input.className).toContain('focus-visible:outline-2');
-    expect(input.className).toContain('focus-visible:outline-primary-4');
+    expect(input.className).toContain('focus-visible:outline-interactive-text');
   });
 });